Finished!
Finally finished this clone!
I tried to make it look and act as much like the actual Monkeytype website as possible, but there are a few behaviours I couldn't implement. While I'm done working on it for now, I will try to improve it based on feedback in the future.